refactor(home-14): add of member's api and model
