+++ src/api/member.api.ts
@@ -1,19 +1,23 @@
 import { Member, mapMember } from "@/shared/models/member.model";
 
-const MEMBERS: Member[] = [
+const CLUBS: Member[] = [
+    // Replace this by the API call to gather all the clubs.
     {
+        //
         nom     : "Nom Prenom",
         photo   : "/fillingPic.gif",
         fonction: "Fonction",
         email   : "nom.prenom@universite-paris-saclay.fr",
     },
     {
+        //
         nom     : "Nom Prenom",
         photo   : "/fillingPic.gif",
         fonction: "Fonction",
         email   : "nom.prenom@universite-paris-saclay.fr",
     },
     {
+        //
         nom     : "Nom Prenom",
         photo   : "/fillingPic.gif",
         fonction: "Fonction",
@@ -24,5 +28,5 @@ const MEMBERS: Member[] = [
 export async function getMembers(): Promise<Member[]> {
     await fetch("https://data.bde-pps.fr/bde/images/logo/bde.svg");
 
-    return MEMBERS.map(mapMember);
+    return CLUBS.map((club) => { return mapMember(club); });
 }
