+++ src/ui/components/members/list.component.tsx
@@ -1,9 +1,8 @@
 import { FC, ReactNode, useEffect, useState } from "react";
 import { getMembers } from "@/api/member.api";
 import { Member } from "@/shared/models/member.model";
-import MemberCard from "@/ui/components/members/card.component";
+import MemberCard from "./card.component";
 import { CircularProgress } from "@mui/material";
-import { unknownToString } from "@/shared/utils/common/convert.util";
 
 const MembersListComp: FC = (): ReactNode => {
     const [members, setmembers] = useState<Member[]>([]);
@@ -11,33 +10,25 @@ const MembersListComp: FC = (): ReactNode => {
     useEffect(() => {
         getMembers()
             .then(setmembers)
-            .catch((error: unknown) => { alert("Can't retrieve Members list :" + unknownToString(error)); });
+            .catch((error: unknown) => { alert("Can't retrieve Members list :" + String(error)); });
     }, []);
 
     return (
         <>
             <div className="w-4/5 m-[2%]">
                 <div className="w-full">
-                    <h1 className="underline text-5xl font-semibold">Les Evenements à venir :</h1>
+                    <h1 className="underline text-5xl font-semibold">Les Membres du Bureau</h1>
                 </div>
 
-                {
-                    members.length === 0
-                        ? <CircularProgress />
-                        : (
-                            <ul className="mt-[3%] columns-1 lg:columns-2">
-                                {
-                                    members.map((member) => {
-                                        return (
-                                            <li>
-                                                <MemberCard member={member} />
-                                            </li>
-                                        );
-                                    })
-                                }
-                            </ul>
-                        )
-                }
+                {members.length === 0
+                    ? <CircularProgress />
+                    : (
+                        <div className="mt-[3%]">
+                            {members.map((club) => {
+                                return <MemberCard key={club.nom} {...club} />;
+                            })}
+                        </div>
+                    )}
             </div>
         </>
     );
